[Refactor] : StudyPost 패키지 분리
