+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/controller/StudyPostController.java
@@ -1,14 +1,13 @@
-package kr.co.studyhubinu.studyhubserver.study.controller;
+package kr.co.studyhubinu.studyhubserver.studypost.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
-import kr.co.studyhubinu.studyhubserver.study.dto.request.CreatePostRequest;
-import kr.co.studyhubinu.studyhubserver.study.dto.request.UpdatePostRequest;
-import kr.co.studyhubinu.studyhubserver.study.dto.response.GetBookmarkedPostsResponse;
-import kr.co.studyhubinu.studyhubserver.study.dto.response.GetMyPostResponse;
-import kr.co.studyhubinu.studyhubserver.study.service.StudyPostService;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.request.CreatePostRequest;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.request.UpdatePostRequest;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.GetBookmarkedPostsResponse;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.GetMyPostResponse;
+import kr.co.studyhubinu.studyhubserver.studypost.service.StudyPostService;
 import kr.co.studyhubinu.studyhubserver.user.dto.data.UserId;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Slice;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
