+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/controller/StudyPostFindController.java
@@ -1,10 +1,12 @@
-package kr.co.studyhubinu.studyhubserver.study.controller;
+package kr.co.studyhubinu.studyhubserver.studypost.controller;
 
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.v3.oas.annotations.Operation;
-import kr.co.studyhubinu.studyhubserver.study.dto.response.*;
-import kr.co.studyhubinu.studyhubserver.study.service.StudyPostService;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.FindPostResponseByAll;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.FindPostResponseByRemainingSeat;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.FindPostResponseByString;
+import kr.co.studyhubinu.studyhubserver.studypost.service.StudyPostService;
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
@@ -48,16 +50,16 @@ public class StudyPostFindController {
     }
 
 
-    //    @Operation(summary = "스터디 게시글 인기순 조회", description = "parameter 칸에 " +
-//            "페이지 정보는 page, 조회할 행 개수는 size 에 입력해주세요")
-//    @ApiImplicitParams({
-//            @ApiImplicitParam(name = "page", value = "페이지", required = true),
-//            @ApiImplicitParam(name = "size", value = "사이즈", required = true)
-//    })
-//    @GetMapping("/hot")
-//    public ResponseEntity<Slice<StudyPostEntity>> findPostByBookMark(@RequestParam int page, @RequestParam int size) {
-//        Pageable pageable = PageRequest.of(page, size);
-//        return ResponseEntity.ok(studyPostService.findPostResponseByBookMark(pageable));
-//    }
+    @Operation(summary = "스터디 게시글 인기순 조회", description = "parameter 칸에 " +
+        "페이지 정보는 page, 조회할 행 개수는 size 에 입력해주세요")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "page", value = "페이지", required = true),
+            @ApiImplicitParam(name = "size", value = "사이즈", required = true)
+    })
+    @GetMapping("/hot")
+    public ResponseEntity<Slice<FindPostResponseByRemainingSeat>> findPostByBookMark(@RequestParam int page, @RequestParam int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        return ResponseEntity.ok(studyPostService.findPostResponseByBookMark(pageable));
+    }
 
 }
