+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/dto/data/GetMyPostData.java
@@ -1,26 +1,23 @@
-package kr.co.studyhubinu.studyhubserver.study.dto.response;
+package kr.co.studyhubinu.studyhubserver.studypost.dto.data;
 
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
 import lombok.Getter;
 
 @Getter
-public class FindPostResponseByMajor {
-
+public class GetMyPostData {
     private Long postId;
     private MajorType major;
     private String title;
     private String content;
-    private int leftover;
-    private int studyPerson;
+    private int remainingSeat;
     private boolean close;
 
-    public FindPostResponseByMajor(Long postId, MajorType major, String title, String content, int leftover, Integer studyPerson, Boolean close) {
+    public GetMyPostData(Long postId, MajorType major, String title, String content, int remainingSeat, boolean close) {
         this.postId = postId;
         this.major = major;
         this.title = title;
         this.content = content;
-        this.leftover = leftover;
-        this.studyPerson = studyPerson;
+        this.remainingSeat = remainingSeat;
         this.close = close;
     }
 }
