+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/dto/data/UpdateStudyPostInfo.java
@@ -1,6 +1,5 @@
-package kr.co.studyhubinu.studyhubserver.study.dto.data;
+package kr.co.studyhubinu.studyhubserver.studypost.dto.data;
 
-import kr.co.studyhubinu.studyhubserver.study.domain.StudyPostEntity;
 import kr.co.studyhubinu.studyhubserver.study.enums.StudyWayType;
 import kr.co.studyhubinu.studyhubserver.user.enums.GenderType;
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
@@ -10,8 +9,8 @@ import lombok.Getter;
 import java.time.LocalDate;
 
 @Getter
-public class StudyPostInfo {
-
+public class UpdateStudyPostInfo {
+    private Long postId;
     private Long userId;
     private String title;
     private String content;
@@ -25,7 +24,8 @@ public class StudyPostInfo {
     private LocalDate studyEndDate;
 
     @Builder
-    public StudyPostInfo(Long userId, String title, String content, String chatUrl, MajorType major, int studyPerson, int penalty, GenderType gender, StudyWayType studyWay, LocalDate studyStartDate, LocalDate studyEndDate) {
+    public UpdateStudyPostInfo(Long postId, Long userId, String title, String content, String chatUrl, MajorType major, int studyPerson, int penalty, GenderType gender, StudyWayType studyWay, LocalDate studyStartDate, LocalDate studyEndDate) {
+        this.postId = postId;
         this.userId = userId;
         this.title = title;
         this.content = content;
@@ -37,24 +37,5 @@ public class StudyPostInfo {
         this.studyWay = studyWay;
         this.studyStartDate = studyStartDate;
         this.studyEndDate = studyEndDate;
-    }
-
-    public StudyPostEntity toEntity(Long userId) {
-
-        return StudyPostEntity.builder()
-                .title(title)
-                .content(content)
-                .chatUrl(chatUrl)
-                .major(major)
-                .studyPerson(studyPerson)
-                .filteredGender(gender)
-                .studyWay(studyWay)
-                .penalty(penalty)
-                .studyStartDate(studyStartDate)
-                .studyEndDate(studyEndDate)
-                .userId(userId)
-                .remainingSeat(studyPerson)
-                .build();
-
     }
 }
