+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/dto/response/FindPostResponseByAll.java
@@ -1,23 +1,26 @@
-package kr.co.studyhubinu.studyhubserver.study.dto.data;
+package kr.co.studyhubinu.studyhubserver.studypost.dto.response;
 
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
 import lombok.Getter;
 
 @Getter
-public class GetMyPostData {
+public class FindPostResponseByAll {
     private Long postId;
     private MajorType major;
     private String title;
     private String content;
-    private int remainingSeat;
+    private int leftover;
+    private int studyPerson;
     private boolean close;
 
-    public GetMyPostData(Long postId, MajorType major, String title, String content, int remainingSeat, boolean close) {
+    public FindPostResponseByAll(Long postId, MajorType major, String title, String content, int leftover, Integer studyPerson, Boolean close) {
         this.postId = postId;
         this.major = major;
         this.title = title;
         this.content = content;
-        this.remainingSeat = remainingSeat;
+        this.leftover = leftover;
+        this.studyPerson = studyPerson;
         this.close = close;
     }
+
 }
