+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/dto/response/FindPostResponseByBookMark.java
@@ -1,10 +1,11 @@
-package kr.co.studyhubinu.studyhubserver.study.dto.response;
+package kr.co.studyhubinu.studyhubserver.studypost.dto.response;
 
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
 import lombok.Getter;
 
 @Getter
-public class FindPostResponseByAll {
+public class FindPostResponseByBookMark {
+
     private Long postId;
     private MajorType major;
     private String title;
@@ -12,8 +13,9 @@ public class FindPostResponseByAll {
     private int leftover;
     private int studyPerson;
     private boolean close;
+    private Long bookMarkCount;
 
-    public FindPostResponseByAll(Long postId, MajorType major, String title, String content, int leftover, Integer studyPerson, Boolean close) {
+    public FindPostResponseByBookMark(Long postId, MajorType major, String title, String content, int leftover, int studyPerson, Boolean close, Long bookMarkCount) {
         this.postId = postId;
         this.major = major;
         this.title = title;
@@ -21,6 +23,6 @@ public class FindPostResponseByAll {
         this.leftover = leftover;
         this.studyPerson = studyPerson;
         this.close = close;
+        this.bookMarkCount = bookMarkCount;
     }
-
 }
