+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/dto/response/FindPostResponseByMajor.java
@@ -1,10 +1,10 @@
-package kr.co.studyhubinu.studyhubserver.study.dto.response;
+package kr.co.studyhubinu.studyhubserver.studypost.dto.response;
 
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
 import lombok.Getter;
 
 @Getter
-public class FindPostResponseByContent {
+public class FindPostResponseByMajor {
 
     private Long postId;
     private MajorType major;
@@ -14,7 +14,7 @@ public class FindPostResponseByContent {
     private int studyPerson;
     private boolean close;
 
-    public FindPostResponseByContent(Long postId, MajorType major, String title, String content, int leftover, Integer studyPerson, Boolean close) {
+    public FindPostResponseByMajor(Long postId, MajorType major, String title, String content, int leftover, Integer studyPerson, Boolean close) {
         this.postId = postId;
         this.major = major;
         this.title = title;
