+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/repository/StudyPostRepository.java
@@ -1,8 +1,7 @@
-package kr.co.studyhubinu.studyhubserver.study.repository;
+package kr.co.studyhubinu.studyhubserver.studypost.repository;
 
-import kr.co.studyhubinu.studyhubserver.study.domain.StudyPostEntity;
-import kr.co.studyhubinu.studyhubserver.study.dto.data.GetMyPostData;
-import kr.co.studyhubinu.studyhubserver.study.dto.response.GetMyPostResponse;
+import kr.co.studyhubinu.studyhubserver.studypost.domain.StudyPostEntity;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.data.GetMyPostData;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface StudyPostRepository extends JpaRepository<StudyPostEntity, Long>, StudyPostRepositoryCustom {
 
-    @Query("SELECT new kr.co.studyhubinu.studyhubserver.study.dto.data.GetMyPostData(sp.id, sp.major, sp.title, sp.content, sp.remainingSeat, sp.close) " +
+    @Query("SELECT new kr.co.studyhubinu.studyhubserver.studypost.dto.data.GetMyPostData(sp.id, sp.major, sp.title, sp.content, sp.remainingSeat, sp.close) " +
             "FROM StudyPostEntity sp " +
             "WHERE sp.postedUserId = :userId")
     Slice<GetMyPostData> findByPostedUserId(Long userId, Pageable pageable);
