+++ src/main/java/kr/co/studyhubinu/studyhubserver/studypost/repository/StudyPostRepositoryCustom.java
@@ -1,9 +1,10 @@
-package kr.co.studyhubinu.studyhubserver.study.repository;
+package kr.co.studyhubinu.studyhubserver.studypost.repository;
 
-import kr.co.studyhubinu.studyhubserver.study.dto.data.GetBookmarkedPostsData;
-import kr.co.studyhubinu.studyhubserver.study.dto.response.*;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.data.GetBookmarkedPostsData;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.FindPostResponseByAll;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.FindPostResponseByRemainingSeat;
+import kr.co.studyhubinu.studyhubserver.studypost.dto.response.FindPostResponseByString;
 import kr.co.studyhubinu.studyhubserver.user.enums.MajorType;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
@@ -15,6 +16,8 @@ public interface StudyPostRepositoryCustom {
     Slice<FindPostResponseByAll> findByAll(Pageable pageable);
 
     Slice<GetBookmarkedPostsData> findPostsByBookmarked(Long userId, Pageable pageable);
+
+    Slice<FindPostResponseByRemainingSeat> findPostsByRemainingSeat(Pageable pageable);
 
     //Slice<StudyPostEntity> findByBookMark(Pageable pageable);
 }
