+++ src/main/java/kr/co/studyhubinu/studyhubserver/user/repository/UserRepositoryCustom.java
@@ -1,9 +1,5 @@
 package kr.co.studyhubinu.studyhubserver.user.repository;
 
-import kr.co.studyhubinu.studyhubserver.study.dto.response.FindPostResponseByAll;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
-
 public interface UserRepositoryCustom {
 
 //    Slice<FindPostResponseByAll> findMyPost(Long userId, Pageable pageable);
